requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
axios==0.4.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.4
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
#django-mongoengine==0.5.4
djangorestframework==3.14.0
#djongo==1.3.6
dnspython==2.4.0
executing==1.2.0
fonttools==4.40.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
idna==2.10
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
lxml==4.9.3
markdown-it-py==3.0.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
#mongoengine==0.27.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.25.0
openai==0.27.8
openapi==1.1.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.1
populartimes==2.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
Pygments==2.15.1
#pymongo==4.4.1
pyparsing==3.0.9
pyscopg2==66.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.1
pyzmq==25.1.0
referencing==0.30.0
requests==2.25.1
rich==13.4.2
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
#sklearn==0.0.post5 - this should be coved by scikit-learn. DM commenting it out
sniffio==1.3.0
#sqlparse==0.3.1
sqlparse
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.4
wcwidth==0.2.6
yarl==1.9.2